fixed distinct generator tests
