+++ tests/distinct_generators/test_text_generator.py
@@ -33,8 +33,8 @@ class TestTextGeneratorFunctions(unittest.TestCase):
             self.assertEqual(text_arr.size, nr)
 
     def test_chars(self):
-        chars_set = {"0", "1"}
-        text_arr = random_text(self.rng, chars=["0", "1"])
+        vocab_set = {"0", "1"}
+        text_arr = random_text(self.rng, vocab=["0", "1"])
         for t in text_arr:
             for char in t:
-                self.assertIn(char, chars_set)
+                self.assertIn(char, vocab_set)
